backend/requirements.txt
fastapi
passlib
uvicorn
boto3
python-dotenv
passlib
pydantic
PyJWT
cryptography
